+++ examples/data-sources/fmc_geolocations/data-source.tf
@@ -1,0 +1,6 @@
+data "fmc_geolocations" "example" {
+  items = {
+    "my_geolocations" = {
+    }
+  }
+}