+++ examples/resources/fmc_geolocations/resource.tf
@@ -1,0 +1,16 @@
+resource "fmc_geolocations" "example" {
+  items = {
+    my_geolocations = {
+      continents = [
+        {
+          id = 6
+        }
+      ]
+      countries = [
+        {
+          id = 616
+        }
+      ]
+    }
+  }
+}